Add static_files_are_immutable parameter
This adds the parameter static_files_are_immutable to responder.API(). When set
to true, Whitenoise is configured such that static files are served with
Cache-Control: "max-age={ten years}, public, immutable".

This is handy for setups where all filenames in /static include the hash of
their content, such as the Webpack setup created by create-react-app.